requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
entrypoints==0.3
flake8==3.7.8
future==0.17.1
idna==2.8
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-telegram-bot==12.1.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
tornado==6.0.3
urllib3==1.25.5
